// packages/shared-types/src/enums.ts
export enum WorkspaceRole {
  OWNER = 'OWNER',
  ADMIN = 'ADMIN',
  MANAGER = 'MANAGER',
  DEVELOPER = 'DEVELOPER',
  QA = 'QA',
  CLIENT = 'CLIENT',
  GUEST = 'GUEST',
}